Extensive rewrite
Needed to re-think some things after reconsidering the spec.
This service was a bit overloaded: Should only be responsible for the
DB-heavy-lifting of defining casts between LOA and retrieving data.

Still need to figure out indexing, which might be something I could
outsource to a translation service? (date-to-month-id, timecaster, only
it's aware of what columns actually map to what information by way of a
metadata table?)