Year-unit resource format
Now data is queried per-year as well as per-column.
This is less efficient than bulk-querying, but is easier to reason with
in terms of caching, so I think the performance hit is justifiable.
Also, it greatly reduces the need for performant hardware, since each
request won't be much bigger than 250k. Scaling horizontally (e.g
requesting lots of stuff at once) might migitate the performance hit of
not requesting in bulk.

Now returns a 404 when no data was found for the particular time-period
(corresponding to the querying format of asking for year-files).