Docs, testing, small tweaks
Wrote tests for the QP, as well as auxiliary functions.
The function API changed, by making LOA metadata (index columns) an
argument. This makes the function a lot easier to test and reason about.

I have some ideas for refactoring, but I'm dropping them in an issue and
moving on to exposing the new functionality through a web api.

Next up, publish!