Started rewrite
Started rewriting BDR with three focus points:

1. Use new external QP library
2. Add LOA metadata handling
3. Fix various bugs.

This first commit contains a major rewrite of database connection
handling, via the addition of a DatabaseLayer class. This cleans up
app.py, which should be as small and readable as possible. It also makes
it really easy to use different databases for handling base data and LOA
metadata (which I plan to do).

In addition, the commit replaces calls to the old join_network code with
the new one, from the library.

It also changes the LOA call to the old "settings" loas, with the new
one that uses the database, using the old defined LOAs as defaults. This
might be deprecated.