Wrote query composer
Wrote a query composer class that encapsulates the querying process in
a series of smaller functions. Also, the rewrite uses PyMonad, making
the process type safe (it always returns either Left or Right).

In addition, the query compose uses the LOA model to get LOA
information, which lets it easily integrate with the new database LOAs.

Not tested on actual data yet, but extensively unittested.